refactor: move activitypub api to events
